Move LifecycleScopeProvider to new lifecycle artifact
