+++ sample/src/main/java/com/uber/autodispose/recipes/AutoDisposeView.java
@@ -23,7 +23,7 @@ import android.support.annotation.RequiresApi;
 import android.util.AttributeSet;
 import android.view.View;
 import com.uber.autodispose.LifecycleEndedException;
-import com.uber.autodispose.LifecycleScopeProvider;
+import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
 import com.uber.autodispose.android.ViewScopeProvider;
 import io.reactivex.Observable;
 import io.reactivex.functions.Function;
