+++ sample/src/main/java/com/uber/autodispose/recipes/AutoDisposeViewHolder.java
@@ -20,7 +20,7 @@ import android.support.v7.widget.BindAwareViewHolder;
 import android.view.View;
 
 import com.uber.autodispose.LifecycleEndedException;
-import com.uber.autodispose.LifecycleScopeProvider;
+import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
 
 import org.jetbrains.annotations.Nullable;
